mcps/postgressql-mcp-master/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axis-postgres-mcp"
version = "1.0.0"
description = "Read-only PostgreSQL MCP server with schema inspection tools."
requires-python = ">=3.11"
dependencies = [
  "asyncpg>=0.31.0",
  "mcp>=1.26.0",
  "pydantic>=2.12.0",
  "python-dotenv>=1.2.0",
  "sqlglot>=27.10.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
  "integration: tests that require a real PostgreSQL instance",
]

[tool.ruff]
line-length = 100
target-version = "py311"
